Swiftify Enums RE #4
Uses the clang AST to pull out the enums.
Hexlifying the int values isn't strictly necessary since the AEKeyword is type aliased to UInt32. AEKeyword (or FourCharCode or UInt32) is necessary to prevent swift from inferring a signed type (or 64bit). And the ScriptingBridge underlying API uses AEKeyword as the preferred type for properties.